+++ lib/src/models/models.dart
@@ -1,4 +1,6 @@
 export 'authenticate_response.dart';
+export 'discoveries/discover_response.dart';
+export 'discoveries/get_rooms_response.dart';
 export 'gateway_administration/log_management_response.dart';
 export 'gateway_administration/get_versions_response.dart';
 export 'gateway_administration/reboot_response.dart';